backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
psycopg2-binary
requests
flask-login
python-dotenv
authlib>=1.0,<2.0
google-auth
passlib[bcrypt]==1.*
itsdangerous==2.*
gunicorn